design das tasks corrigido.
