+++ src/main/java/app/controllers/TitlebarController.java
@@ -1,0 +1,12 @@
+package app.controllers;
+
+import javafx.scene.control.Label;
+import javafx.scene.input.MouseEvent;
+
+public class TitlebarController {
+    public Label btnCloseApp;
+
+    public void btnCloseApp(MouseEvent event) {
+        System.exit(0);
+    }
+}
