+++ src/main/java/app/models/Category.java
@@ -1,10 +1,10 @@
 package app.models;
 
 public abstract class Category {
-    public String type;
-    public String icon;
-    public String description;
-    public int id;
+    private String type;
+    private String icon;
+    private String description;
+    private int id;
 
 
     public Category(String type, String icon, String description, int id) {
