+++ src/main/resources/app/styles/login.css
@@ -13,6 +13,7 @@
     -fx-background-color: white;
     -fx-pref-width: 100;
     -fx-alignment: center;
+    -fx-cursor: HAND;
 }
 .signin{
     -fx-text-fill: black;
