+++ src/main/resources/app/views/closeBar.fxml
@@ -6,8 +6,9 @@
 <?import javafx.scene.control.*?>
 <?import javafx.scene.layout.*?>
 
-<AnchorPane xmlns="http://javafx.com/javafx" xmlns:fx="http://javafx.com/fxml" fx:controller="app.controllers.MainController" prefHeight="20.0" prefWidth="800.0" style="-fx-background-color: #DBC3F1FF;">
-    <children>
-        <Label layoutX="777.0" layoutY="2.0" onMouseClicked="#btnCloseApp" prefHeight="13.0" prefWidth="15.0" style="-fx-background-radius: 20 20 20 20; -fx-background-color: red; -fx-border-radius: 20 20 20 20; -fx-alignment: center;" text="X" textAlignment="CENTER" />
-    </children>
+<AnchorPane xmlns="http://javafx.com/javafx" xmlns:fx="http://javafx.com/fxml"
+            fx:controller="app.controllers.TitlebarController" prefHeight="20.0" prefWidth="800.0"
+            styleClass="closeBar" stylesheets="@../styles/style.css">
+    <Label layoutX="777.0" layoutY="2.0" onMouseClicked="#btnCloseApp" prefHeight="13.0" prefWidth="17.0"
+           styleClass="closeButton" textAlignment="CENTER"/>
 </AnchorPane>
