--- src/main/resources/app/views/tasksViews.fxml
@@ -1,111 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.ComboBox?>
-<?import javafx.scene.control.DatePicker?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.control.ScrollPane?>
-<?import javafx.scene.control.TextField?>
-<?import javafx.scene.effect.DropShadow?>
-<?import javafx.scene.layout.AnchorPane?>
-<?import javafx.scene.layout.ColumnConstraints?>
-<?import javafx.scene.layout.GridPane?>
-<?import javafx.scene.layout.Region?>
-<?import javafx.scene.layout.RowConstraints?>
-<?import javafx.scene.text.Font?>
-<?import javafx.scene.text.Text?>
-
-<AnchorPane fx:id="tasks" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="600.0" prefWidth="800.0" styleClass="bg" stylesheets="@../styles/tasks.css" xmlns="http://javafx.com/javafx/16" xmlns:fx="http://javafx.com/fxml/1" fx:controller="app.controllers.TasksController">
-   <Button fx:id="addNewTask" layoutX="704.0" layoutY="517.0" mnemonicParsing="false" onAction="#onClickBtnAddTask" styleClass="addTaskBtn" text="+" textFill="WHITE">
-      <font>
-         <Font name="Consolas Bold" size="24.0" />
-      </font>
-   </Button>
-   <AnchorPane fx:id="taskCreator" layoutX="171.0" layoutY="115.0" prefHeight="370.0" prefWidth="458.0" style="-fx-background-color: #dac9fd; -fx-background-radius: 50px; -fx-border-radius: 50px 55px;" visible="false">
-      <effect>
-         <DropShadow />
-      </effect>
-      <GridPane layoutX="34.0" layoutY="14.0" prefHeight="329.0" prefWidth="391.0">
-         <columnConstraints>
-            <ColumnConstraints hgrow="SOMETIMES" maxWidth="334.0000163180505" minWidth="11.056566694021683" prefWidth="11.056566694021683" />
-            <ColumnConstraints hgrow="SOMETIMES" maxWidth="364.2642082955842" minWidth="106.67921553680048" prefWidth="175.83016029105352" />
-            <ColumnConstraints hgrow="SOMETIMES" maxWidth="283.2830662467635" minWidth="10.0" prefWidth="129.64155147847194" />
-            <ColumnConstraints hgrow="SOMETIMES" maxWidth="283.2830662467635" minWidth="10.0" prefWidth="112.30188958455608" />
-         </columnConstraints>
-         <rowConstraints>
-            <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
-            <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
-            <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
-            <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
-            <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
-            <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
-            <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
-            <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
-         </rowConstraints>
-         <TextField fx:id="descriptionId" style="fx-background-radius: 50px;" GridPane.columnSpan="4" GridPane.rowIndex="5">
-            <font>
-               <Font size="20.0" />
-            </font>
-         </TextField>
-         <Text strokeType="OUTSIDE" strokeWidth="0.0" textAlignment="RIGHT" wrappingWidth="155.80675506591797" GridPane.columnIndex="1" GridPane.halignment="LEFT" GridPane.rowIndex="7" GridPane.valignment="CENTER">
-            <font>
-               <Font name="Segoe MDL2 Assets" size="20.0" />
-            </font>
-         </Text>
-         <TextField id="taskname" fx:id="taskId" maxHeight="-Infinity" maxWidth="-Infinity" prefHeight="40.0" prefWidth="392.0" style="fx-background-radius: 50px;" GridPane.columnSpan="4" GridPane.rowIndex="1">
-            <font>
-               <Font size="20.0" />
-            </font>
-         </TextField>
-         <Text strokeType="OUTSIDE" strokeWidth="0.0" text="Título da Tarefa" wrappingWidth="161.751953125" GridPane.columnIndex="1" GridPane.valignment="CENTER">
-            <font>
-               <Font name="Segoe MDL2 Assets" size="20.0" />
-            </font>
-         </Text>
-         <Text strokeType="OUTSIDE" strokeWidth="0.0" text="Categoria" wrappingWidth="128.916015625" GridPane.columnIndex="1" GridPane.rowIndex="2" GridPane.valignment="CENTER">
-            <font>
-               <Font name="Segoe MDL2 Assets" size="20.0" />
-            </font>
-         </Text>
-         <Text strokeType="OUTSIDE" strokeWidth="0.0" text="Data Limite" textAlignment="CENTER" wrappingWidth="203.98501586914062" GridPane.columnIndex="1" GridPane.halignment="LEFT" GridPane.rowIndex="6" GridPane.valignment="CENTER">
-            <font>
-               <Font name="Segoe MDL2 Assets" size="20.0" />
-            </font>
-         </Text>
-         <DatePicker fx:id="dateId" prefHeight="34.0" prefWidth="118.0" promptText="dd/MM/aaaa" style="-fx-border-radius: 50px; -fx-background-color: white;" GridPane.columnIndex="2" GridPane.halignment="RIGHT" GridPane.rowIndex="6" GridPane.valignment="CENTER" />
-         <Button fx:id="savingTask" alignment="CENTER" contentDisplay="CENTER" defaultButton="true" graphicTextGap="0.0" maxHeight="51.0" maxWidth="113.0" minHeight="41.0" minWidth="40.0" mnemonicParsing="false" onAction="#saveTask" prefHeight="41.0" prefWidth="40.0" styleClass="saveTaskButton" text="OK" textAlignment="JUSTIFY" textFill="WHITE" textOverrun="CENTER_ELLIPSIS" GridPane.columnIndex="3" GridPane.halignment="CENTER" GridPane.rowIndex="6" GridPane.rowSpan="2" GridPane.valignment="CENTER">
-            <font>
-               <Font name="System Bold" size="15.0" />
-            </font>
-            <effect>
-               <DropShadow />
-            </effect>
-         </Button>
-         <Text strokeType="OUTSIDE" strokeWidth="0.0" text="Descrição" textOrigin="CENTER" wrappingWidth="291.1444625854492" GridPane.columnIndex="1" GridPane.rowIndex="4" GridPane.valignment="CENTER">
-            <font>
-               <Font name="Segoe MDL2 Assets" size="20.0" />
-            </font>
-         </Text>
-         <Label prefHeight="24.0" prefWidth="128.0" text="Horário Limite" textAlignment="RIGHT" GridPane.columnIndex="1" GridPane.halignment="RIGHT" GridPane.rowIndex="7">
-            <font>
-               <Font name="Segoe MDL2 Assets" size="20.0" />
-            </font>
-         </Label>
-         <TextField fx:id="time" alignment="CENTER" maxHeight="-Infinity" maxWidth="-Infinity" prefHeight="0.0" prefWidth="118.0" promptText="HH:MM:SS" GridPane.columnIndex="2" GridPane.rowIndex="7">
-            <font>
-               <Font size="20.0" />
-            </font>
-         </TextField>
-         <ComboBox fx:id="categoryId" prefHeight="49.0" prefWidth="392.0" style="-fx-background-color: white;" GridPane.columnSpan="4" GridPane.rowIndex="3" />
-         <Button alignment="BOTTOM_RIGHT" mnemonicParsing="false" onAction="#closeTaskTab" prefHeight="48.0" prefWidth="100.0" style="-fx-background-color: transparent;" styleClass="button" text="X" textAlignment="RIGHT" textFill="WHITE" GridPane.columnIndex="3" GridPane.halignment="RIGHT" GridPane.valignment="CENTER">
-            <font>
-               <Font name="System Bold" size="22.0" />
-            </font>
-         </Button>
-      </GridPane>
-   </AnchorPane>
-   <ScrollPane layoutX="160.0" layoutY="78.0" prefHeight="522.0" prefWidth="531.0" style="-fx-background-color: white;" stylesheets="@../styles/tasks.css" />
-   <AnchorPane prefHeight="600.0" prefWidth="42.0" style="-fx-background-color: white;" />
-   <Label layoutX="217.0" layoutY="38.0" prefHeight="29.0" prefWidth="475.0" styleClass="title" text="Title of the category" />
-   <Region fx:id="titleIcon" layoutX="160.0" layoutY="28.0" prefHeight="50.0" prefWidth="50.0" />
-</AnchorPane>
